+++ airplane-mode.php
@@ -1,15 +1,16 @@
 <?php
 /**
  * Plugin Name: Airplane Mode
- * Plugin URI: http://reaktivstudios.com/
+ * Plugin URI: https://github.com/norcross/airplane-mode
  * Description: Control loading of external files when developing locally
  * Author: Andrew Norcross
- * Author URI: http://reaktivstudios.com/
- * Version: 0.1.7
+ * Author URI: http://andrewnorcross.com/
+ * Version: 0.2.2
  * Text Domain: airplane-mode
  * Requires WP: 4.4
  * Domain Path: languages
  * GitHub Plugin URI: https://github.com/norcross/airplane-mode
+ * @package airplane-mode
  */
 
 /*
@@ -48,7 +49,12 @@ if ( ! defined( 'AIRMDE_DIR' ) ) {
 
 // Set our version if not already defined.
 if ( ! defined( 'AIRMDE_VER' ) ) {
-	define( 'AIRMDE_VER', '0.1.7' );
+	define( 'AIRMDE_VER', '0.2.2' );
+}
+
+// Load our WP-CLI helper if that is defined and available.
+if ( defined( 'WP_CLI' ) && WP_CLI ) {
+	require_once dirname( __FILE__ ) . '/inc/wp-cli.php';
 }
 
 // Ensure the class has not already been loaded.
@@ -104,7 +110,6 @@ if ( ! class_exists( 'Airplane_Mode_Core' ) ) {
 			add_action( 'admin_bar_menu',                       array( $this, 'admin_bar_toggle'        ),  9999    );
 			add_action( 'wp_enqueue_scripts',                   array( $this, 'toggle_css'              ),  9999    );
 			add_action( 'admin_enqueue_scripts',                array( $this, 'toggle_css'              ),  9999    );
-			add_action( 'login_enqueue_scripts',                array( $this, 'toggle_css'              ),  9999    );
 
 			// Body class on each location for the display.
 			add_filter( 'body_class',                           array( $this, 'body_class'              )           );
@@ -123,7 +128,8 @@ if ( ! class_exists( 'Airplane_Mode_Core' ) ) {
 			add_filter( 'install_plugins_tabs',                 array( $this, 'plugin_add_tabs'         )           );
 
 			// Theme update API for different calls.
-			add_filter( 'themes_api_args',                      array( $this, 'bypass_theme_api'        ),  10, 2   );
+			add_filter( 'themes_api',                           array( $this, 'bypass_theme_api_call'   ),  10, 3   );
+			add_filter( 'themes_api_result',                    array( $this, 'bypass_theme_api_result' ),  10, 3   );
 
 			// Time based transient checks.
 			add_filter( 'pre_site_transient_update_themes',     array( $this, 'last_checked_themes'     )           );
@@ -132,6 +138,9 @@ if ( ! class_exists( 'Airplane_Mode_Core' ) ) {
 			add_filter( 'site_transient_update_themes',         array( $this, 'remove_update_array'     )           );
 			add_filter( 'site_transient_update_plugins',        array( $this, 'remove_update_array'     )           );
 
+			// Disable fetching languages from online
+			add_filter( 'site_transient_available_translations', array( $this, 'available_translations' ), 9999, 1  );
+
 			// Our activation / deactivation triggers.
 			register_activation_hook( __FILE__,                 array( $this, 'create_setting'          )           );
 			register_deactivation_hook( __FILE__,               array( $this, 'remove_setting'          )           );
@@ -139,14 +148,17 @@ if ( ! class_exists( 'Airplane_Mode_Core' ) ) {
 			// All our various filter checks.
 			if ( $this->enabled() ) {
 
-				// Keep jetpack from attempting external requests.
-				add_filter( 'jetpack_development_mode',             '__return_true', 9999 );
+				// Allows locally defined JETPACK_DEV_DEBUG constant to override filter.
+				if ( ! defined( 'JETPACK_DEV_DEBUG' ) ) {
+					// Keep jetpack from attempting external requests.
+					add_filter( 'jetpack_development_mode',         '__return_true', 9999 );
+				}
+
+				// Prevent BuddyPress from falling back to Gravatar avatars.
+				add_filter( 'bp_core_fetch_avatar_no_grav',         '__return_true' );
 
 				// Disable automatic updater updates.
 				add_filter( 'automatic_updater_disabled',           '__return_true' );
-
-				// Hijack the themes api setup to bypass the API call.
-				add_filter( 'themes_api',                           '__return_true' );
 
 				// Tell WordPress we are on a version control system to add additional blocks.
 				add_filter( 'automatic_updates_is_vcs_checkout',    '__return_true' );
@@ -233,7 +245,6 @@ if ( ! class_exists( 'Airplane_Mode_Core' ) ) {
 		 */
 		public function create_setting() {
 			add_site_option( 'airplane-mode', 'on' );
-			set_transient( 'available_translations', '', 999999999999 );
 			set_transient( 'wporg_theme_feature_list', array(), 999999999999 );
 		}
 
@@ -243,7 +254,6 @@ if ( ! class_exists( 'Airplane_Mode_Core' ) ) {
 		public function remove_setting() {
 			delete_option( 'airplane-mode' );
 			delete_site_option( 'airplane-mode' );
-			delete_transient( 'available_translations' );
 			delete_transient( 'wporg_theme_feature_list' );
 		}
 
@@ -357,11 +367,11 @@ if ( ! class_exists( 'Airplane_Mode_Core' ) ) {
 		public function body_class( $classes ) {
 
 			// Add the class based on the current status.
-			$classes[]	= $this->enabled() ? 'airplane-mode-enabled' : 'airplane-mode-disabled';
+			$classes[]  = $this->enabled() ? 'airplane-mode-enabled' : 'airplane-mode-disabled';
 
 			// Also add in the margin setup for Query Monitor because I'm a perfectionist.
 			if ( ! class_exists( 'QueryMonitor' ) || defined( 'QM_DISABLED' ) && QM_DISABLED ) {
-				$classes[]	= 'airplane-mode-no-qm';
+				$classes[]  = 'airplane-mode-no-qm';
 			}
 
 			// Return our array of classes.
@@ -491,9 +501,10 @@ if ( ! class_exists( 'Airplane_Mode_Core' ) ) {
 
 			$url_host = parse_url( $url, PHP_URL_HOST );
 
-			// Allow the request to pass through if the URL host matches the site's host:
+			// Allow the request to pass through if the URL host matches the site's host.
 			if ( $url_host && parse_url( home_url(), PHP_URL_HOST ) === $url_host ) {
-				// But allow this to be disabled via a filter:
+
+				// But allow this to be disabled via a filter.
 				if ( apply_filters( 'airplane_mode_allow_local_bypass', true, $url, $args ) ) {
 					return $status;
 				}
@@ -508,18 +519,65 @@ if ( ! class_exists( 'Airplane_Mode_Core' ) ) {
 		 */
 		public function toggle_css() {
 
+			// Don't display CSS on the front-end if the admin bar is not loading.
+			if ( ! is_admin() && ! is_admin_bar_showing() ) {
+				return;
+			}
+
 			// Set a suffix for loading the minified or normal.
-			$suffix = defined( 'SCRIPT_DEBUG' ) && SCRIPT_DEBUG ? '.css' : '.min.css';
+			$file   = defined( 'SCRIPT_DEBUG' ) && SCRIPT_DEBUG ? 'airplane-mode.css' : 'airplane-mode.min.css';
+
+			// Set a version for browser caching.
+			$vers   = defined( 'SCRIPT_DEBUG' ) && SCRIPT_DEBUG ? time() : AIRMDE_VER;
 
 			// Load the CSS file itself.
-			wp_enqueue_style( 'airplane-mode', plugins_url( '/lib/css/airplane-mode' . $suffix, __FILE__ ), array(), AIRMDE_VER, 'all' );
+			wp_enqueue_style( 'airplane-mode', plugins_url( '/lib/css/' . $file, __FILE__ ), array(), $vers, 'all' );
+		}
+
+		/**
+		 * Sets the mode.
+		 *
+		 * @param  string $mode Desired mode ('on' or 'off').
+		 *
+		 * @return bool Whether the setting changed.
+		 */
+		public function set_mode( $mode = 'on' ) {
+			if ( ! in_array( $mode, array( 'on', 'off' ) ) ) {
+				$mode = 'on';
+			}
+
+			// Update the setting.
+			$return = update_site_option( 'airplane-mode', $mode );
+
+			// Fire action to allow for functions to run on status change.
+			do_action( 'airplane_mode_status_change', $mode );
+
+			return $return;
+		}
+
+		/**
+		 * Enables airplane mode.
+		 *
+		 * @return bool Whether the setting changed.
+		 */
+		public function enable() {
+			return self::set_mode( 'on' );
+		}
+
+		/**
+		 * Disables airplane mode.
+		 *
+		 * @return bool Whether the setting changed.
+		 */
+		public function disable() {
+			return self::set_mode( 'off' );
 		}
 
 		/**
 		 * Check the user action from the toggle switch to set the option
 		 * to 'on' or 'off'.
 		 *
-		 * @return void if any of the sanity checks fail and we bail early.
+		 * @return void If any of the sanity checks fail and we bail early.
 		 */
 		public function toggle_check() {
 
@@ -545,11 +603,7 @@ if ( ! class_exists( 'Airplane_Mode_Core' ) ) {
 			// Delete old per-site option.
 			delete_option( 'airplane-mode' );
 
-			// Update the setting.
-			update_site_option( 'airplane-mode', sanitize_key( $_REQUEST['airplane-mode'] ) );
-
-			// Fire action to allow for functions to run on status change.
-			do_action( 'airplane_mode_status_change', $switch );
+			self::set_mode( $switch );
 
 			// And go about our business.
 			wp_redirect( self::get_redirect() );
@@ -690,6 +744,7 @@ if ( ! class_exists( 'Airplane_Mode_Core' ) ) {
 				delete_site_transient( 'update_core' );
 				delete_site_transient( 'update_plugins' );
 				delete_site_transient( 'update_themes' );
+				delete_site_transient( 'wporg_theme_feature_list' );
 			}
 		}
 
@@ -757,23 +812,45 @@ if ( ! class_exists( 'Airplane_Mode_Core' ) ) {
 		}
 
 		/**
-		 * Hijack the themes api setup to bypass the API call.
+		 * Override the API call made for pulling themes from the .org repo.
 		 *
-		 * @param object $args    Arguments used to query for installer pages from the Themes API.
-		 * @param string $action  Requested action. Likely values are 'theme_information',
-		 *                        'feature_list', or 'query_themes'.
+		 * @param false|object|array $override  Whether to override the WordPress.org Themes API. Default false.
+		 * @param string             $action    Requested action. Likely values are 'theme_information',
+		 *                                      'feature_list', or 'query_themes'.
+		 * @param object             $args      Arguments used to query for installer pages from the Themes API.
 		 *
-		 * @return bool           true or false depending on the type of query
+		 * @return bool                         True if enabled, otherwise the existng value.
 		 */
-		public function bypass_theme_api( $args, $action ) {
+		public function bypass_theme_api_call( $override, $action, $args ) {
 
 			// Bail if disabled.
 			if ( ! $this->enabled() ) {
-				return $args;
+				return $override;
 			}
 
 			// Return false on feature list to avoid the API call.
-			return ! empty( $action ) && 'feature_list' === $action ? false : $args;
+			return ! empty( $action ) && 'feature_list' === $action ? true : $override;
+		}
+
+		/**
+		 * Hijack the expected themes API result.
+		 *
+		 * @param array|object|WP_Error $res     WordPress.org Themes API response.
+		 * @param string                $action  Requested action. Likely values are 'theme_information',
+		 *                                       'feature_list', or 'query_themes'.
+		 * @param object                $args    Arguments used to query for installer pages from the WordPress.org Themes API.
+		 *
+		 * @return bool                          An empty array if enabled, otherwise the existng result.
+		 */
+		public function bypass_theme_api_result( $res, $action, $args ) {
+
+			// Bail if disabled.
+			if ( ! $this->enabled() ) {
+				return $res;
+			}
+
+			// Return false on feature list to avoid the API call.
+			return ! empty( $action ) && in_array( $action, array( 'feature_list', 'query_themes' ) ) ? array() : $res;
 		}
 
 		/**
@@ -793,9 +870,9 @@ if ( ! class_exists( 'Airplane_Mode_Core' ) ) {
 
 			// Return our object.
 			return (object) array(
-				'last_checked'		=> time(),
-				'updates'			=> array(),
-				'version_checked'	=> $wp_version,
+				'last_checked'      => time(),
+				'updates'           => array(),
+				'version_checked'   => $wp_version,
 			);
 		}
 
@@ -824,10 +901,10 @@ if ( ! class_exists( 'Airplane_Mode_Core' ) ) {
 
 			// Return our object.
 			return (object) array(
-				'last_checked'		=> time(),
-				'updates'			=> array(),
-				'version_checked'	=> $wp_version,
-				'checked'			=> $data,
+				'last_checked'      => time(),
+				'updates'           => array(),
+				'version_checked'   => $wp_version,
+				'checked'           => $data,
 			);
 		}
 
@@ -861,11 +938,179 @@ if ( ! class_exists( 'Airplane_Mode_Core' ) ) {
 
 			// Return our object.
 			return (object) array(
-				'last_checked'		=> time(),
-				'updates'			=> array(),
-				'version_checked'	=> $wp_version,
-				'checked'			=> $data,
+				'last_checked'      => time(),
+				'updates'           => array(),
+				'version_checked'   => $wp_version,
+				'checked'           => $data,
 			);
+		}
+
+		/**
+		 * Filter for languages list transient. Returns locally available translations
+		 * to avoid request into wp.org translation API.
+		 *
+		 * @param mixed $translations  Translation data returned from transient API.
+		 *
+		 * @return array                List of available languages.
+		 */
+		public function available_translations( $translations ) {
+
+			// Bail if disabled.
+			if ( ! $this->enabled() ) {
+				return $translations;
+			}
+
+			/**
+			 * If transient still contains list of languages just use those.
+			 * Otherwise fallback to mimicked data which we create here.
+			 */
+			if ( $translations && ! empty( $translations ) ) {
+				return $translations;
+			} else {
+				$installed_languages = get_available_languages();
+				return $this->get_offline_languages( $installed_languages );
+			}
+		}
+
+		/**
+		 * Returns list of languages installed locally with full mimicked meta data.
+		 *
+		 * @param array $installed_languages  List of locally available languages.
+		 *
+		 * @return array                List of available languages in offline mode.
+		 */
+		private function get_offline_languages( $installed_languages ) {
+
+			// This is list of languages which are available from translations API.
+			$offline_languages = $this->get_offline_translation_information();
+
+			// Call the global WP version.
+			global $wp_version;
+
+			// Tell WordPress that all translations are recent even though they can be old.
+			$date = date_i18n( 'Y-m-d H:is' , time() );
+
+			// Set an empty array of the available languages.
+			$available_languages = array();
+
+			// Loop through our installed languages.
+			foreach ( $installed_languages as $lang ) {
+
+				// Try to mimick the data that WordPress puts into 'available_translations' transient.
+				$settings = array(
+					'language'  => $lang,
+					'iso'       => array( $lang ),
+					'version'   => $wp_version,
+					'updated'   => $date,
+					'strings'   => array(
+						'continue' => __( 'Continue' ),
+					),
+					'package'   => "https://downloads.wordpress.org/translation/core/{$wp_version}/{$lang}.zip",
+				);
+
+				// Combine the mimicked data with data we have stored in $offline_languages to give realistic output.
+				$available_languages[ $lang ] = array_merge( $settings, $offline_languages[ $lang ] );
+			}
+
+			// And return our language sets.
+			return $available_languages;
+		}
+
+		/**
+		 * We can't use wp_get_available_translations() offine.
+		 * This function tries to return something similiar.
+		 *
+		 * @return array     List of wordpress language meta data.
+		 */
+		private function get_offline_translation_information() {
+
+			// Build out the list of languages to use.
+			$languages  = array(
+				'ar'                => array( 'english_name' => 'Arabic', 'native_name' => 'العربية' ),
+				'ary'               => array( 'english_name' => 'Moroccan Arabic', 'native_name' => 'العربية المغربية' ),
+				'az'                => array( 'english_name' => 'Azerbaijani', 'native_name' => 'Azərbaycan dili' ),
+				'azb'               => array( 'english_name' => 'South Azerbaijani', 'native_name' => 'گؤنئی آذربایجان' ),
+				'bg_BG'             => array( 'english_name' => 'Bulgarian', 'native_name' => 'Български' ),
+				'bn_BD'             => array( 'english_name' => 'Bengali', 'native_name' => 'বাংলা' ),
+				'bs_BA'             => array( 'english_name' => 'Bosnian', 'native_name' => 'Bosanski' ),
+				'ca'                => array( 'english_name' => 'Catalan', 'native_name' => 'Català' ),
+				'ceb'               => array( 'english_name' => 'Cebuano', 'native_name' => 'Cebuano' ),
+				'cs_CZ'             => array( 'english_name' => 'Czech', 'native_name' => 'Čeština‎' ),
+				'cy'                => array( 'english_name' => 'Welsh', 'native_name' => 'Cymraeg' ),
+				'da_DK'             => array( 'english_name' => 'Danish', 'native_name' => 'Dansk' ),
+				'de_DE_formal'      => array( 'english_name' => 'German (Formal)', 'native_name' => 'Deutsch (Sie)' ),
+				'de_DE'             => array( 'english_name' => 'German', 'native_name' => 'Deutsch' ),
+				'de_CH_informal'    => array( 'english_name' => '(Switzerland, Informal)', 'native_name' => 'Deutsch (Schweiz, Du)' ),
+				'de_CH'             => array( 'english_name' => 'German (Switzerland)', 'native_name' => 'Deutsch (Schweiz)' ),
+				'el'                => array( 'english_name' => 'Greek', 'native_name' => 'Ελληνικά' ),
+				'en_CA'             => array( 'english_name' => 'English (Canada)', 'native_name' => 'English (Canada)' ),
+				'en_ZA'             => array( 'english_name' => 'English (South Africa)', 'native_name' => 'English (South Africa)' ),
+				'en_AU'             => array( 'english_name' => 'English (Australia)', 'native_name' => 'English (Australia)' ),
+				'en_NZ'             => array( 'english_name' => 'English (New Zealand)', 'native_name' => 'English (New Zealand)' ),
+				'en_GB'             => array( 'english_name' => 'English (UK)', 'native_name' => 'English (UK)' ),
+				'eo'                => array( 'english_name' => 'Esperanto', 'native_name' => 'Esperanto' ),
+				'es_CL'             => array( 'english_name' => 'Spanish (Chile)', 'native_name' => 'Español de Chile' ),
+				'es_AR'             => array( 'english_name' => 'Spanish (Argentina)', 'native_name' => 'Español de Argentina' ),
+				'es_PE'             => array( 'english_name' => 'Spanish (Peru)', 'native_name' => 'Español de Perú' ),
+				'es_MX'             => array( 'english_name' => 'Spanish (Mexico)', 'native_name' => 'Español de México' ),
+				'es_CO'             => array( 'english_name' => 'Spanish (Colombia)', 'native_name' => 'Español de Colombia' ),
+				'es_ES'             => array( 'english_name' => 'Spanish (Spain)', 'native_name' => 'Español' ),
+				'es_VE'             => array( 'english_name' => 'Spanish (Venezuela)', 'native_name' => 'Español de Venezuela' ),
+				'es_GT'             => array( 'english_name' => 'Spanish (Guatemala)', 'native_name' => 'Español de Guatemala' ),
+				'et'                => array( 'english_name' => 'Estonian', 'native_name' => 'Eesti' ),
+				'eu'                => array( 'english_name' => 'Basque', 'native_name' => 'Euskara' ),
+				'fa_IR'             => array( 'english_name' => 'Persian', 'native_name' => 'فارسی' ),
+				'fi'                => array( 'english_name' => 'Finnish', 'native_name' => 'Suomi' ),
+				'fr_BE'             => array( 'english_name' => 'French (Belgium)', 'native_name' => 'Français de Belgique' ),
+				'fr_FR'             => array( 'english_name' => 'French (France)', 'native_name' => 'Français' ),
+				'fr_CA'             => array( 'english_name' => 'French (Canada)', 'native_name' => 'Français du Canada' ),
+				'gd'                => array( 'english_name' => 'Scottish Gaelic', 'native_name' => 'Gàidhlig' ),
+				'gl_ES'             => array( 'english_name' => 'Galician', 'native_name' => 'Galego' ),
+				'haz'               => array( 'english_name' => 'Hazaragi', 'native_name' => 'هزاره گی' ),
+				'he_IL'             => array( 'english_name' => 'Hebrew', 'native_name' => 'עִבְרִית' ),
+				'hi_IN'             => array( 'english_name' => 'Hindi', 'native_name' => 'हिन्दी' ),
+				'hr'                => array( 'english_name' => 'Croatian', 'native_name' => 'Hrvatski' ),
+				'hu_HU'             => array( 'english_name' => 'Hungarian', 'native_name' => 'Magyar' ),
+				'hy'                => array( 'english_name' => 'Armenian', 'native_name' => 'Հայերեն' ),
+				'id_ID'             => array( 'english_name' => 'Indonesian', 'native_name' => 'Bahasa Indonesia' ),
+				'is_IS'             => array( 'english_name' => 'Icelandic', 'native_name' => 'Íslenska' ),
+				'it_IT'             => array( 'english_name' => 'Italian', 'native_name' => 'Italiano' ),
+				'ja'                => array( 'english_name' => 'Japanese', 'native_name' => '日本語' ),
+				'ka_GE'             => array( 'english_name' => 'Georgian', 'native_name' => 'ქართული' ),
+				'ko_KR'             => array( 'english_name' => 'Korean', 'native_name' => '한국어' ),
+				'lt_LT'             => array( 'english_name' => 'Lithuanian', 'native_name' => 'Lietuvių kalba' ),
+				'mk_MK'             => array( 'english_name' => 'Macedonian', 'native_name' => 'Македонски јазик' ),
+				'mr'                => array( 'english_name' => 'Marathi', 'native_name' => 'मराठी' ),
+				'ms_MY'             => array( 'english_name' => 'Malay', 'native_name' => 'Bahasa Melayu' ),
+				'my_MM'             => array( 'english_name' => 'Myanmar (Burmese)', 'native_name' => 'ဗမာစာ' ),
+				'nb_NO'             => array( 'english_name' => 'Norwegian (Bokmål)', 'native_name' => 'Norsk bokmål' ),
+				'nl_NL'             => array( 'english_name' => 'Dutch', 'native_name' => 'Nederlands' ),
+				'nl_NL_formal'      => array( 'english_name' => 'Dutch (Formal)', 'native_name' => 'Nederlands (Formeel)' ),
+				'nn_NO'             => array( 'english_name' => 'Norwegian (Nynorsk)', 'native_name' => 'Norsk nynorsk' ),
+				'oci'               => array( 'english_name' => 'Occitan', 'native_name' => 'Occitan' ),
+				'pl_PL'             => array( 'english_name' => 'Polish', 'native_name' => 'Polski' ),
+				'ps'                => array( 'english_name' => 'Pashto', 'native_name' => 'پښتو' ),
+				'pt_BR'             => array( 'english_name' => 'Portuguese (Brazil)', 'native_name' => 'Português do Brasil' ),
+				'pt_PT'             => array( 'english_name' => 'Portuguese (Portugal)', 'native_name' => 'Português' ),
+				'ro_RO'             => array( 'english_name' => 'Romanian', 'native_name' => 'Română' ),
+				'ru_RU'             => array( 'english_name' => 'Russian', 'native_name' => 'Русский' ),
+				'sk_SK'             => array( 'english_name' => 'Slovak', 'native_name' => 'Slovenčina' ),
+				'sl_SI'             => array( 'english_name' => 'Slovenian', 'native_name' => 'Slovenščina' ),
+				'sq'                => array( 'english_name' => 'Albanian', 'native_name' => 'Shqip' ),
+				'sr_RS'             => array( 'english_name' => 'Serbian', 'native_name' => 'Српски језик' ),
+				'sv_SE'             => array( 'english_name' => 'Swedish', 'native_name' => 'Svenska' ),
+				'th'                => array( 'english_name' => 'Thai', 'native_name' => 'ไทย' ),
+				'tl'                => array( 'english_name' => 'Tagalog', 'native_name' => 'Tagalog' ),
+				'tr_TR'             => array( 'english_name' => 'Turkish', 'native_name' => 'Türkçe' ),
+				'ug_CN'             => array( 'english_name' => 'Uighur', 'native_name' => 'Uyƣurqə' ),
+				'uk'                => array( 'english_name' => 'Ukrainian', 'native_name' => 'Українська' ),
+				'vi'                => array( 'english_name' => 'Vietnamese', 'native_name' => 'Tiếng Việt' ),
+				'zh_CN'             => array( 'english_name' => 'Chinese (China)', 'native_name' => '简体中文' ),
+				'zh_TW'             => array( 'english_name' => 'Chinese (Taiwan)', 'native_name' => '繁體中文' ),
+			);
+
+			// Allow adding or removing languages.
+			return apply_filters( 'airplane_mode_offline_languages', $languages );
 		}
 
 		/**
@@ -945,7 +1190,8 @@ if ( ! class_exists( 'Airplane_Mode_Core' ) ) {
 			$this->http_count++;
 		}
 
-	} // end class
+		// End class.
+	}
 
 } //end class_exists