+++ CHANGES.md
@@ -1,3 +1,20 @@
+#### Version 0.2.2 - 2016/12/07
+* Adding WP-CLI support. props @markjaquith
+
+#### Version 0.2.1 - 2016/10/19
+* Fix settings page language drowdown when used offline. props @onnimonni
+
+#### Version 0.2.0 - 2016/08/26
+* modify CSS loading for front-end, removed loading for login page. props @barryceelen
+* fixed WP.org theme API call request to avoid `WP_Error` return. props @onnimonni
+
+#### Version 0.1.9 - 2016/07/25
+* prevent BuddyPress from falling back to Gravatar. props @johnbillion
+
+#### Version 0.1.8 - 2016/07/12
+* allow `JETPACK_DEV_DEBUG` constant to take priority over filter. props @kopepasah
+* added additional CSS for upcoming 4.6. change to upload tab.
+
 #### Version 0.1.7 - 2016/05/18
 * allow local HTTP calls with optional filter. props @johnbillion
 * add back index.php link to main dashboard menu item